requirements.txt
customtkinter
watchpoints
requests
CTkListbox
CTkMessagebox
asynctkinter
requests-toolbelt
pillow